algoverse-AI/requirements.txt
# --- Core ---
numpy==1.26.4
tqdm==4.66.4

# --- Transcription ---
openai-whisper==20231117
ffmpeg-python==0.2.0

# --- Embeddings ---
sentence-transformers==2.6.1

# --- Vector DB ---
chromadb==1.3.7

# --- API ---
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4

# --- LLM Clients ---
openai==1.14.3
requests==2.31.0 

# --- Utils ---
python-dotenv==1.0.1

youtube-transcript-api==0.6.2
yt-dlp>=2024.12.13

